# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api of lmon",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "All requests",
                "description": "All requests",
                "operationId": "all_applicants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "identification": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "amount": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "quotas": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "total": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Store request",
                "description": "Store request",
                "operationId": "store_applicants",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "identification": {
                                        "type": "numeric",
                                        "example": "40237252669"
                                    },
                                    "amount": {
                                        "type": "numeric",
                                        "example": "10000"
                                    },
                                    "quotas": {
                                        "type": "numeric",
                                        "example": "4"
                                    },
                                    "total": {
                                        "type": "numeric",
                                        "example": "10400"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/requests/user/{id}": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "See request",
                "description": "See request",
                "operationId": "watch_request_by_user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "identification": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "amount": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "quotas": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "total": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "See request",
                "description": "See request",
                "operationId": "watch_applicants",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "identification": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "amount": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "quotas": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "total": {
                                            "type": "numeric",
                                            "example": ""
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "Update request",
                "description": "Update request",
                "operationId": "update_applicants",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "identification": {
                                        "type": "numeric",
                                        "example": "40237252669"
                                    },
                                    "amount": {
                                        "type": "numeric",
                                        "example": "10000"
                                    },
                                    "quotas": {
                                        "type": "numeric",
                                        "example": "4"
                                    },
                                    "total": {
                                        "type": "numeric",
                                        "example": "10400"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Requests"
                ],
                "summary": "Delete request",
                "description": "Delete request",
                "operationId": "delete_applicants",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}